requirements.txt
diffusers
transformers
scipy
ftfy
python-dotenv